Consolidate openid related support in openid directory
