+++ lib/galaxy/openid/openid_conf.xml
