+++ lib/galaxy/openid/openid_manager.py
